rename package and module name springcloud 2 feign.
