+++ shenyu-examples/shenyu-examples-sdk/pom.xml
@@ -67,7 +67,7 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.shenyu</groupId>
-                <artifactId>shenyu-spring-boot-starter-sdk-springcloud</artifactId>
+                <artifactId>shenyu-spring-boot-starter-sdk-feign</artifactId>
                 <version>${shenyu-sdk.version}</version>
             </dependency>
         </dependencies>
