+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sdk/shenyu-spring-boot-starter-sdk-feign/pom.xml
@@ -18,15 +18,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.apache.shenyu</groupId>
-        <artifactId>shenyu-spring-boot-starter</artifactId>
+        <artifactId>shenyu-spring-boot-starter-sdk</artifactId>
         <version>2.6.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>shenyu-spring-boot-starter-sdk</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>shenyu-spring-boot-starter-sdk-feign</artifactId>
 
-    <modules>
-        <module>shenyu-spring-boot-starter-sdk-spring</module>
-        <module>shenyu-spring-boot-starter-sdk-feign</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shenyu</groupId>
+            <artifactId>shenyu-sdk-feign</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
 </project>
